pyproject.toml
[project]
name = "spider-snoop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "boto3>=1.42.24",
    "clamd>=1.0.2",
    "email-validator>=2.3.0",
    "faker>=40.1.0",
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "moviepy>=1.0.3",
    "oletools>=0.60.2",
    "openai>=2.14.0",
    "openpyxl>=3.1.5",
    "passlib>=1.7.4",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.5.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.21",
    "reportlab>=4.4.7",
    "sentence-transformers>=5.2.2",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.45",
    "torch>=2.10.0",
    "transformers>=5.1.0",
    "uvicorn>=0.40.0",
    "werkzeug>=3.1.5",
    "yara-python>=4.5.4",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "requests>=2.32.5",
    "starlette>=0.50.0",
]
